requirements.txt
Django==4.0.4
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djongo==1.3.6
drf-yasg==1.21.5
greenlet==2.0.2
gevent==22.10.2

protobuf~=3.19.4
grpcio~=1.47.0
grpcio-tools~=1.44.0

requests==2.28.1
urllib3~=1.26.15
pymongo~=3.12.1
PyMySQL~=1.0.3
SQLAlchemy~=2.0.11

pandas~=2.0.0
openpyxl~=3.1.2
xlsxwriter~=3.1.2
PyYAML~=6.0
jsonlines~=3.1.0
